broke out config into a sub-module
