+++ src/config/mod.rs
@@ -1,0 +1,8 @@
+//! all logic related to instantiating a running configuration
+
+mod container;
+mod utils;
+#[cfg(test)]
+mod tests;
+
+pub use container::Configuration;
